+++ systems/__init__.py
@@ -1,3 +1,1 @@
-from .NJP00 import NJP00
-from .PRL00 import PRL00
 from .Mod00 import Mod00